pyproject.toml
[tool.poetry]
name = "bactopia"
version = "2.0.0"
description = "A Python package for working with Bactopia"
authors = [
    "Robert A. Petit III <robbie.petit@gmail.com>",
]
license = "MIT"
readme = "README.md"
homepage = "https://bactopia.github.io/"
repository = "https://github.com/bactopia/bactopia-py"
keywords = ["bioinformatics", "bacteria", "bactopia", "SRA", "ENA"]

[tool.poetry.scripts]
bactopia-citations = "bactopia.cli.citations:main"
bactopia-datasets = "bactopia.cli.datasets:main"
bactopia-download = "bactopia.cli.download:main"
bactopia-prepare = "bactopia.cli.prepare:main"
bactopia-search = "bactopia.cli.search:main"
bactopia-summary = "bactopia.cli.summary:main"
bactopia-update = "bactopia.cli.update:main"
bactopia-status = "bactopia.cli.status:main"
bactopia-workflows = "bactopia.cli.workflows:main"
bactopia-atb-formatter = "bactopia.cli.atb.atb_formatter:main"
bactopia-atb-downloader = "bactopia.cli.atb.atb_downloader:main"
bactopia-merge-schemas = "bactopia.cli.helpers.merge_schemas:main"
bactopia-pubmlst-setup = "bactopia.cli.pubmlst.setup:main"
bactopia-prune = "bactopia.cli.prune:main"
bactopia-pubmlst-build = "bactopia.cli.pubmlst.build:main"
bactopia-test = "bactopia.cli.testing:main"
bactopia-lint = "bactopia.cli.lint:main"
bactopia-catalog = "bactopia.cli.catalog:main"
bactopia-review-tests = "bactopia.cli.review:main"
# Pipeline utility scripts (called from Nextflow module shell blocks)
bactopia-check-fastqs = "bactopia.cli.pipeline.check_fastqs:main"
bactopia-check-assembly-accession = "bactopia.cli.pipeline.check_assembly_accession:main"
bactopia-cleanup-coverage = "bactopia.cli.pipeline.cleanup_coverage:main"
bactopia-mask-consensus = "bactopia.cli.pipeline.mask_consensus:main"
bactopia-kraken-bracken-summary = "bactopia.cli.pipeline.kraken_bracken_summary:main"
bactopia-scrubber-summary = "bactopia.cli.pipeline.scrubber_summary:main"
bactopia-midas-summary = "bactopia.cli.pipeline.midas_summary:main"
bactopia-teton-prepare = "bactopia.cli.pipeline.teton_prepare:main"
bactopia-bracken-to-excel = "bactopia.cli.pipeline.bracken_to_excel:main"

[tool.poetry.dependencies]
python = "^3.9.0"
requests = "^2.28.2"
rich = "^13.3.1"
rich-click = "^1.6.1"
pandas = "^2.2.0"
tqdm = "^4.66.5"
rauth = "^0.7.3"
jinja2 = "^3.1.6"
pyyaml = "^6.0"
biopython = "^1.80"
openpyxl = "^3.1.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^8.0"
pytest-cov = "^6.0"
pytest-mock = "^3.14"
responses = "^0.25"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E203", "E501", "E713", "F401", "E402", "F403"]

[tool.ruff.lint.isort]
known-first-party = ["bactopia"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: tests requiring BACTOPIA_RESULTS data (skip with -m 'not integration')",
]
addopts = ["--strict-markers", "-ra", "--tb=short"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
